feat: ajustando estrutura de arquivos
