+++ database/db.py
@@ -1,0 +1,14 @@
+from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
+DATABASE_URL = "sqlite:///loja_de_carros.db"
+
+engine = create_engine(DATABASE_URL)
+
+Base = declarative_base()
+
+Session = sessionmaker(bind=engine)
+
+def init_db():
+    Base.metadata.create_all(bind=engine)
