--- main.py
@@ -1,21 +1,0 @@
-from entidades import Veiculo_service , Cliente_service , Venda_service
-
-# veiculo = Veiculo(marca=marca,modelo=modelo,ano=ano,preco=preco,cor=cor,numero_de_rodas=numero_de_rodas,placa=placa , Is_available=Is_available)
-
-# Veiculo_service.cadastro_veiculo('punto','fiat','1990','23000','azul','4','QKJD-2456' , '0')
-
-
-# Cliente_service.cadastro_cliente('Igor' , '04194240' , '112345678' , '123@gmail' )
-
-Cliente_service.listar_cliente()
-
-# Veiculo_service.lista_veiculos()
-
-# Cliente_service.update_cliente(1,'Jonatan' , '04194240' , '112345678' , '123@gmail')
-
-# Cliente_service.delete_cliente(1)
-
-# Venda_service.cadastro_venda('1' , '2' , '2024-11-03 16:42:56.766940' , '23000')
-# Venda_service.cadastro_venda('1' , '2' , '2024-11-03 16:42:56.766940' , '30000')
-
-# Venda_service.lista_vendas()
